Finish up the notification
