--- YCForum/Functions/Notifi/HiddenNotificationTrigger.swift
@@ -1,11 +1,0 @@
-//
-//  HiddenNotificationTrigger.swift
-//  YCForum
-//
-//  Created by Sonic Liu on 11/12/2023.
-//
-
-import Foundation
-
-
-
