+++ YCForum/Views/MainPage.swift
@@ -130,6 +130,7 @@ struct Mainpage: View {
             DispatchQueue.main.asyncAfter(deadline: .now() + 1.0) {
                 notificationManag.requestAuthorization()
             }
+            notificationManag.askingNewWork()
             notificationManag.sendDailyNotificationForUnfinishedWork(info)
             saveSettings.loadSettings()
             print(saveSettings.selectedSubject)
